% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-opinions_2019.R
\docType{data}
\name{opinions_2019}
\alias{opinions_2019}
\title{Supreme Court Opinions from 2019 Session}
\format{
A tibble with 210 rows and 4 variables:
\describe{
\item{text}{Text of the opinion}
\item{author}{Author of the opinion}
\item{type}{Type of the opinion}
\item{case}{Name of the case}
}
}
\source{
\url{https://www.supremecourt.gov/}
}
\usage{
opinions_2019
}
\description{
A dataset containing the text from slip opinions from the US Supreme Court
  2019 session.
}
\keyword{datasets}
